checks to see if the method 'klass' is 'char'
If statement to figure out the method. If it's 'char' it's adds a class
char - item value i.e.  If the letter is 'a' it's adds a class of
'char-a'.
This allows you create kerning pairs. You can create kerning pair rules
for all the character combos. You won't need to know the content
beforehand. Use case could be on a CMS where you won't know the
headings beforehand but want strict kerning rules.